Add file lock to avoid parallel backup
Now uses the envfile for flock, to avoid running same
backup at same time. This avoids problems with pre- and post-
backup scripts clashing if running parallel backups by mistake,
as they are not protected by the internal locking in borg.

If the lock is already taken, the second backup simply fails.